requirements.txt
streamlit
tensorflow
Pillow
numpy
scikit-learn
matplotlib
seaborn
